Add files via upload
The project is about selecting a night of the week that most people can attend. 
Create functions (Python) that help automate selection process and draft an invitation email with the selected night. Function 'add_gamer' checks availability, then a function that calculates availability based on the frequency table. 
Practice for creating functions and using dictionaries in Python.